+++ src/main/java/de/srendi/advancedperipherals/common/addons/computercraft/peripheral/InventoryManagerPeripheral.java
@@ -2,7 +2,8 @@ package de.srendi.advancedperipherals.common.addons.computercraft.peripheral;
 
 import dan200.computercraft.api.lua.LuaException;
 import dan200.computercraft.api.lua.LuaFunction;
-import dan200.computercraft.shared.util.NBTUtil;
+import dan200.computercraft.client.SoundManager;
+import de.srendi.advancedperipherals.AdvancedPeripherals;
 import de.srendi.advancedperipherals.common.blocks.tileentity.InventoryManagerTile;
 import de.srendi.advancedperipherals.common.configuration.AdvancedPeripheralsConfig;
 import de.srendi.advancedperipherals.common.util.InventoryUtil;
@@ -10,11 +11,15 @@ import de.srendi.advancedperipherals.common.util.ItemUtil;
 import de.srendi.advancedperipherals.common.util.LuaConverter;
 import de.srendi.advancedperipherals.lib.peripherals.BasePeripheral;
 import de.srendi.advancedperipherals.lib.peripherals.owner.TileEntityPeripheralOwner;
+import net.minecraft.block.DispenserBlock;
+import net.minecraft.client.audio.Sound;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ArmorItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.tileentity.DispenserTileEntity;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.Direction;
 import net.minecraftforge.items.CapabilityItemHandler;
@@ -23,6 +28,7 @@ import net.minecraftforge.items.ItemHandlerHelper;
 import net.minecraftforge.items.wrapper.PlayerInvWrapper;
 import net.minecraftforge.registries.ForgeRegistries;
 
+import javax.annotation.Nonnull;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -41,7 +47,7 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
     }
 
     @LuaFunction
-    public final String getOwner() {
+    public final String getOwner() throws LuaException {
         if (getOwnerPlayer() == null)
             return null;
         return getOwnerPlayer().getName().getString();
@@ -49,7 +55,6 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
 
     @LuaFunction(mainThread = true, value = {"pullItems", "addItemToPlayer"})
     public final int addItemToPlayer(String invDirection, int count, Optional<Integer> slot, Optional<String> item) throws LuaException {
-        ensurePlayerIsLinked();
         ItemStack stack = ItemStack.EMPTY;
         if (item.isPresent()) {
             Item item1 = ItemUtil.getRegistryEntry(item.get(), ForgeRegistries.ITEMS);
@@ -65,10 +70,9 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
 
         int invSlot = slot.orElse(0);
 
-        if (invSlot >= inventoryTo.getSlots() || invSlot < 0)
-            throw new LuaException("Inventory out of bounds " + invSlot + " (max: " + (inventoryTo.getSlots() - 1) + ")");
+        //if (invSlot >= inventoryTo.getSlots() || invSlot < 0)
+          //  throw new LuaException("Inventory out of bounds " + invSlot + " (max: " + (inventoryTo.getSlots() - 1) + ")");
 
-        //inventoryTo is checked via ensurePlayerIsLinked()
         if (inventoryFrom == null)
             return 0;
 
@@ -83,7 +87,21 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
             }
 
             if (ItemHandlerHelper.canItemStacksStack(stack, inventoryFrom.getStackInSlot(i))) {
-                int inserted = InventoryUtil.moveItem(inventoryFrom, i, inventoryTo, invSlot, amount);
+                int inserted;
+                if(invSlot >= 100 && invSlot < 104) {
+                    if(!(stack.getItem() instanceof ArmorItem))
+                        throw new LuaException(stack + "is not an armor item. Can't put it into the slot " + invSlot);
+                    //When there is already an item in the slot, just continue
+                    if(!getOwnerPlayer().inventory.armor.get(getArmorSlot(invSlot)).isEmpty())
+                        continue;
+                    getOwnerPlayer().inventory.armor.set(getArmorSlot(invSlot), stack);
+                    inventoryFrom.extractItem(i, 1, false);
+                    //Armor can't be stacked, so we set this just to one
+                    transferableAmount = 1;
+                    //Continue as we don't want to run the normal code for non armor items
+                    continue;
+                }
+                inserted = InventoryUtil.moveItem(inventoryFrom, i, inventoryTo, invSlot, amount);
                 transferableAmount += inserted;
                 amount -= inserted;
 
@@ -98,7 +116,6 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
 
     @LuaFunction(mainThread = true, value = {"pushItems", "removeItemFromPlayer"})
     public final int removeItemFromPlayer(String invDirection, int count, Optional<Integer> slot, Optional<String> item) throws LuaException {
-        ensurePlayerIsLinked();
         ItemStack stack = ItemStack.EMPTY;
         if (item.isPresent()) {
             Item item1 = ItemUtil.getRegistryEntry(item.get(), ForgeRegistries.ITEMS);
@@ -184,59 +201,36 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
         return transferableAmount;
     }
 
+    @Nonnull
     @LuaFunction(value = {"list", "getItems"}, mainThread = true)
     public final Map<Integer, Object> getItems() throws LuaException {
-        ensurePlayerIsLinked();
         Map<Integer, Object> items = new HashMap<>();
-        int i = 0;
+        int i = 0; //Used to let users easily sort the items by the slots. Also a better way for the user to see where a item actually is
         for (ItemStack stack : getOwnerPlayer().inventory.items) {
             if (!stack.isEmpty()) {
-                Map<String, Object> map = new HashMap<>();
-                String displayName = stack.getDisplayName().getString();
-                CompoundNBT nbt = stack.getTag(); //use getTag instead of getOrCreateTag to fix https://github.com/Seniorendi/AdvancedPeripherals/issues/177
-                map.put("name", stack.getItem().getRegistryName().toString());
-                map.put("amount", stack.getCount());
-                map.put("displayName", displayName);
-                if (nbt == null) {
-                    nbt = new CompoundNBT();//ensure compatibility with lua programs relying on a non-nil value
-                }
-                map.put("nbt", NBTUtil.toLua(nbt));
-                map.put("tags", LuaConverter.tagsToList(stack.getItem().getTags()));
-                items.put(i, map);
-                i++;
+                items.put(i, LuaConverter.stackToObject(stack));
             }
+            i++;
         }
         return items;
     }
 
     @LuaFunction(mainThread = true)
     public final Map<Integer, Object> getArmor() throws LuaException {
-        ensurePlayerIsLinked();
         Map<Integer, Object> items = new HashMap<>();
         int i = 0;
         for (ItemStack stack : getOwnerPlayer().inventory.armor) {
+            AdvancedPeripherals.debug("DEBUG1 " + i);
             if (!stack.isEmpty()) {
-                Map<String, Object> map = new HashMap<>();
-                String displayName = stack.getDisplayName().getString();
-                CompoundNBT nbt = stack.getTag(); //use getTag instead of getOrCreateTag to fix https://github.com/Seniorendi/AdvancedPeripherals/issues/177
-                map.put("name", stack.getItem().getRegistryName().toString());
-                map.put("amount", stack.getCount());
-                map.put("displayName", displayName);
-                if (nbt == null) {
-                    nbt = new CompoundNBT();//ensure compatibility with lua programs relying on a non-nil value
-                }
-                map.put("nbt", NBTUtil.toLua(nbt));
-                map.put("tags", LuaConverter.tagsToList(stack.getItem().getTags()));
-                items.put(i, map);
-                i++;
+                items.put(ArmorSlot.getSlotForItem(stack), LuaConverter.stackToObject(stack));
             }
+            i++;
         }
         return items;
     }
 
     @LuaFunction(mainThread = true)
     public final boolean isPlayerEquipped() throws LuaException {
-        ensurePlayerIsLinked();
         for (ItemStack stack : getOwnerPlayer().inventory.armor) {
             if (!stack.isEmpty()) {
                 return true;
@@ -247,7 +241,6 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
 
     @LuaFunction(mainThread = true)
     public final boolean isWearing(int index) throws LuaException {
-        ensurePlayerIsLinked();
         int i = 0;
         for (ItemStack stack : getOwnerPlayer().inventory.armor) {
             if (!stack.isEmpty()) {
@@ -259,13 +252,37 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
         return false;
     }
 
-    private PlayerEntity getOwnerPlayer() {
-        return owner.getOwner();
+
+    @LuaFunction(mainThread = true)
+    public final int getEmptySpace() throws LuaException {
+        int i = 0;
+        for (ItemStack stack : getOwnerPlayer().inventory.items) {
+            if (stack.isEmpty())
+                i++;
+
+        }
+        return i;
     }
 
-    private void ensurePlayerIsLinked() throws LuaException {
-        if (getOwnerPlayer() == null)
+    @LuaFunction(mainThread = true)
+    public final boolean isSpaceAvailable() throws LuaException {
+        return getEmptySpace() > 0;
+    }
+
+    @LuaFunction(mainThread = true)
+    public final int getFreeSlot() throws LuaException {
+        return getOwnerPlayer().inventory.getFreeSlot();
+    }
+
+    @LuaFunction(mainThread = true)
+    public final Map<String, Object> getItemInHand() throws LuaException {
+        return LuaConverter.stackToObject(getOwnerPlayer().inventory.getSelected());
+    }
+
+    private PlayerEntity getOwnerPlayer() throws LuaException {
+        if (owner.getOwner() == null)
             throw new LuaException("The Inventory Manager doesn't have a memory card or it isn't bound to a player.");
+        return owner.getOwner();
     }
 
     private ItemStack insertItem(IItemHandler inventoryTo, ItemStack stack) {
@@ -278,5 +295,57 @@ public class InventoryManagerPeripheral extends BasePeripheral<TileEntityPeriphe
             stack = inventoryTo.insertItem(i, stack, false);
         }
         return stack;
+    }
+
+    private static int getArmorSlot(int index) {
+        switch (index) {
+            case 103:
+                return 3;
+            case 102:
+                return 2;
+            case 101:
+                return 1;
+            case 100:
+                return 0;
+        }
+        return 0;
+    }
+
+    /**
+     * Used to get the proper slot number for armor. See https://docs.srendi.de/ for the slot numbers
+     *
+     * @see InventoryManagerPeripheral#getArmor()
+     */
+    private enum ArmorSlot {
+
+        HELMET_SLOT(103, EquipmentSlotType.HEAD),
+        CHEST_SLOT(102, EquipmentSlotType.CHEST),
+        LEGGINGS_SLOT(101, EquipmentSlotType.LEGS),
+        BOOTS_SLOT(100, EquipmentSlotType.FEET);
+
+        private final int slot;
+        private final EquipmentSlotType slotType;
+
+        ArmorSlot(int slot, EquipmentSlotType slotType) {
+            this.slot = slot;
+            this.slotType = slotType;
+        }
+
+        public int getSlot() {
+            return slot;
+        }
+
+        public static int getSlotForItem(ItemStack stack) {
+            if (stack.getItem() instanceof ArmorItem) {
+                for (ArmorSlot slot : values()) {
+                    if (((ArmorItem) stack.getItem()).getSlot() == slot.slotType) {
+                        return slot.slot;
+                    }
+                }
+            }
+            AdvancedPeripherals.LOGGER.warn("Tried to get armor item slot for non armor item " + stack + ". Returning 0");
+            return 0;
+        }
+
     }
 }