+++ CHANGELOG.md
@@ -10,6 +10,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - FormattedMessage formatting mistake
+- [#203]Fixed Inventory Manager api, slot argument behaviour is now correct.
+
+### Added
+
+- Added noteblock integration
+- Added `getItemInHand`, `getFreeSlot`, `isSpaceAvailable` and `getEmptySpace` to the inventory manager.
+- Added support for armor items. You can use the slots 100-103 to access armor items.
 
 ## [0.7.3r] - 2021-10-13